# packages/typespec/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "www-template API",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/v1/account/settings": {
      "get": {
        "operationId": "getAccountSettings",
        "summary": "Gets the current account settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSettingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["account-settings"],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "operationId": "updateAccountSettings",
        "summary": "Updates the current account settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSettingResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["account-settings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountSettingRequest"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Revokes the current bearer session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["app-auth"],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/passkeys": {
      "get": {
        "operationId": "listPasskeys",
        "summary": "Lists registered passkeys for the current account",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["app-auth"],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountLocale": {
        "type": "string",
        "enum": ["ja", "en"],
        "description": "Product Account が表示と通知に使用する保存済みロケール。運用者向け設定や画面一時状態ではなく、AccountSetting.locale の値だけを表す。"
      },
      "AccountSetting": {
        "type": "object",
        "required": ["locale"],
        "properties": {
          "locale": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountLocale"
              }
            ],
            "description": "現在の Product Account に保存されている表示・通知ロケール。対応値は AccountLocale の ja または en のみ。"
          }
        },
        "description": "Product Account に属する現在の設定。AccountSetting は Account の表示・通知設定を表し、Auth 情報や運用者向け設定を含まない。"
      },
      "AccountSettingResponse": {
        "type": "object",
        "required": ["requestId", "setting"],
        "properties": {
          "requestId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UlidId"
              }
            ],
            "description": "AccountSetting 取得または更新 request に対応する canonical ULID の追跡 ID。"
          },
          "setting": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountSetting"
              }
            ],
            "description": "現在の Product Account に属する保存済み AccountSetting。"
          }
        },
        "description": "現在の Product Account の AccountSetting を返すレスポンス。requestId は追跡用であり、setting には Account 本人の保存済み設定だけを含める。"
      },
...
      "AuthFailureClassification": {
        "type": "string",
        "enum": ["unauthenticated", "session-expired", "account-suspended", "internal-error"]
      },
      "AuthFailureResponse": {
        "type": "object",
        "required": ["requestId", "error"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "error": {
            "$ref": "#/components/schemas/AuthFailureClassification"
          }
        }
      },
      "AuthOperationErrorResponse": {
        "type": "object",
        "required": ["requestId", "error"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "LogoutResponse": {
        "type": "object",
        "required": ["requestId", "revoked"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "revoked": {
            "type": "boolean",
            "enum": [true]
          }
        }
      },
...
      "PasskeyItem": {
        "type": "object",
        "required": ["id", "identifier", "createdAt"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UlidId"
          },
          "identifier": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PasskeyListResponse": {
        "type": "object",
        "required": ["requestId", "passkeys"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "passkeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PasskeyItem"
            }
          }
        }
      },
...
      "UlidId": {
        "type": "string",
        "description": "Canonical ULID string used for auth-owned resource and correlation identifiers."
      },
      "UpdateAccountSettingRequest": {
        "type": "object",
        "required": ["locale"],
        "properties": {
          "locale": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountLocale"
              }
            ],
            "description": "保存したい Product AccountSetting.locale。対応値以外は API 実装で拒否し、永続値を変更しない。"
          }
        },
        "description": "現在の Product Account の AccountSetting.locale を更新するリクエスト。Account ID は bearer session から確定するため body には含めない。"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT 形式のアクセストークンを使用した Bearer 認証。Authorization ヘッダーに `Bearer <JWT>` を付与すること。トークンは短命の JWT であり、accountID・sessionID・iat・exp を含む。"
      }
    }
  },
  "servers": [
    {
      "url": "https://product.www-template.example",
      "description": "Product API server",
      "variables": {}
    }
  ]
}